Remove `test_inference_gpu.py` and integrate its functionality into `test_inference.py`; bump dependency to `rfdetr>=1.5.0`.
