+++ pyproject.toml
@@ -35,7 +35,7 @@ classifiers = [
 keywords = ["machine-learning", "deep-learning", "vision", "ML", "DL", "AI", "DETR", "RF-DETR", "Roboflow"]
 
 dependencies = [
-    "rfdetr>=1.4.3,<2",
+    "rfdetr>=1.5.0rc0,<2",
 ]
 
 [project.optional-dependencies]
